Rename cil scripts
